a minor refactor to make clear the distinction between the iterating code and display code
